Weather Particles Effects and Dropdown Menu
I added the Rain and Snow particle effects and it works using the weather buttons.
I also added the UI - Modern Pack to the build to utilize the UI assets. I used it for the Dropdown Menu but it currently needs some tweaking but its in the scene, the other dropdown is just disabled in scene.